Project_Armor/requirements.txt
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.1.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
scikit-multilearn>=0.2.0
scipy>=1.9.0
tensorboard>=2.10.0
tqdm==4.66.1
PyYAML==6.0.1
albumentations==1.3.1
ultralytics==8.0.200